+++ apps/web/.eslintrc.cjs
@@ -1,4 +1,3 @@
-/** @type {import("eslint").Linter.Config} */
 module.exports = {
   root: true,
   extends: ["@jamsr-ui/eslint-config/next.js"],