+++ apps/web/src/app/layout.tsx
@@ -22,6 +22,8 @@ export const viewport: Viewport = {
   themeColor: "#000000",
 };
 
+let a = 2;
+
 export default function RootLayout({
   children,
 }: Readonly<{