+++ packages/components/accordion/src/test.ts
@@ -1,0 +1,7 @@
+import { useUIStyle } from "@jamsr-ui/styles";
+
+let a = 2;
+
+export const useTest = () => {
+  const {} = useUIStyle();
+};