+++ packages/components/card/src/types.d.ts
@@ -3,11 +3,11 @@ import { CardContentProps } from "./card-content";
 import { CardFooterProps } from "./card-footer";
 import { CardHeaderProps } from "./card-header";
 
-declare module "@jamsr-ui/styles" {
-  export interface UIStylesType {
-    card?: Pick<CardProps, "className">;
-    cardContent?: Pick<CardContentProps, "className">;
-    cardFooter?: Pick<CardFooterProps, "className">;
-    cardHeader?: Pick<CardHeaderProps, "className" | "classNames">;
-  }
-}
+// declare module "@jamsr-ui/styles" {
+//   export interface UIStylesType {
+//     card?: Pick<CardProps, "className">;
+//     cardContent?: Pick<CardContentProps, "className">;
+//     cardFooter?: Pick<CardFooterProps, "className">;
+//     cardHeader?: Pick<CardHeaderProps, "className" | "classNames">;
+//   }
+// }