+++ packages/components/input/src/types.d.ts
@@ -1,7 +1,0 @@
-import { InputProps } from "./input";
-
-declare module "@jamsr-ui/styles" {
-  export interface UIStylesType {
-    input?: Pick<InputProps, "className" | "classNames">;
-  }
-}