+++ packages/core/src/index.ts
@@ -1,2 +1,1 @@
 export { UIProvider } from "./ui-provider";
-export { type UIGlobalStylesType } from "./styles";