+++ packages/styles/.eslintrc.cjs
@@ -1,6 +1,6 @@
 module.exports = {
   root: true,
-  extends: ["@jamsr-ui/eslint-config/next.js"],
+  extends: ["@jamsr-ui/eslint-config/react-library.js"],
   parserOptions: {
     project: true,
   },