+++ packages/styles/tsup.config.ts
@@ -7,9 +7,4 @@ export default defineConfig({
   clean: true,
   minify: true,
   splitting: true,
-  esbuildOptions(options) {
-    options.banner = {
-      js: '"use client"',
-    };
-  },
 });