+++ packages/components/autocomplete/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/autocomplete
 
+## 0.0.73
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/input@0.0.68
+
 ## 0.0.72
 
 ### Patch Changes