+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/core
 
+## 0.0.32
+
+### Patch Changes
+
+- fix types
+
 ## 0.0.31
 
 ### Patch Changes